# Remove style tag from Svelte implementation