@JReinhold 2024-08-28
Does that also make the reference to `./base.css` irrelevant?

https://github.com/platformatic/ssr-performance-showdown/blob/main/svelte/client/index.html#L6

(might not make much of an SSR perf difference)